pyproject.toml
[project]
name = "hermes-discord-skill-audit"
version = "0.1.0"
description = "Hermes Discord Skill Audit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.3.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q"
pythonpath = ["."]
[dependency-groups]
dev = [
    "black>=26.3.1",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
